refactor(rsc): align PPR framework conventions
