+++ packages/plugin-rsc/examples/ppr/src/framework/entry.browser.tsx
@@ -5,9 +5,9 @@ import {
 import React from 'react'
 import { createRoot, hydrateRoot } from 'react-dom/client'
 import { rscStream } from 'rsc-html-stream/client'
+import type { RscPayload } from './entry.rsc'
 import { GlobalErrorBoundary } from './error-boundary'
 import { createRscRenderRequest } from './request'
-import type { RscPayload } from './shared'
 
 async function main() {
   // stash `setPayload` function to trigger re-rendering
