+++ packages/plugin-rsc/examples/ppr/src/framework/entry.rsc.tsx
@@ -7,8 +7,12 @@ import type { PrerenderResult } from 'react-dom/static'
 import { Root } from '../root'
 import { runPrerender } from './ppr-context'
 import { parseRenderRequest } from './request'
-import type { PprData, RscPayload } from './shared'
+import type { PprData } from './shared'
 import { stringToStream } from './stream-utils'
+
+export type RscPayload = {
+  root: React.ReactNode
+}
 
 let manifestPromise: Promise<Record<string, PprData>> | undefined
 
@@ -16,16 +20,18 @@ export function getStaticPaths(): string[] {
   return ['/']
 }
 
-export default async function handler(request: Request): Promise<Response> {
+export default { fetch: handler }
+
+async function handler(request: Request): Promise<Response> {
   const renderRequest = parseRenderRequest(request)
 
   if (renderRequest.isRsc) {
-    const payload: RscPayload = {
+    const rscPayload: RscPayload = {
       root: (
         <Root url={renderRequest.url} timestamp={new Date().toISOString()} />
       ),
     }
-    return new Response(renderToReadableStream(payload), {
+    return new Response(renderToReadableStream(rscPayload), {
       headers: { 'content-type': 'text/x-component;charset=utf-8' },
     })
   }
@@ -33,23 +39,18 @@ export default async function handler(request: Request): Promise<Response> {
   const pprData = import.meta.env.DEV
     ? await handlePpr(renderRequest.request)
     : await loadPprData(renderRequest.url.pathname)
-  const payload: RscPayload = {
+  const rscPayload: RscPayload = {
     root: <Root url={renderRequest.url} timestamp={pprData.staticTimestamp} />,
   }
-  const rscStream = renderToReadableStream(payload)
-  const [rscForSsr, rscForBrowser] = rscStream.tee()
-  const ssr = await import.meta.viteRsc.loadModule<
+  const rscStream = renderToReadableStream(rscPayload)
+  const ssrEntryModule = await import.meta.viteRsc.loadModule<
     typeof import('./entry.ssr')
   >('ssr', 'index')
   const prerenderResult: PrerenderResult = {
     prelude: stringToStream(pprData.html),
     postponed: JSON.parse(pprData.postponed),
   }
-  const htmlStream = await ssr.resumeHtml(
-    rscForSsr,
-    rscForBrowser,
-    prerenderResult,
-  )
+  const htmlStream = await ssrEntryModule.resumeHtml(rscStream, prerenderResult)
 
   return new Response(htmlStream, {
     headers: { 'content-type': 'text/html;charset=utf-8' },
@@ -58,22 +59,22 @@ export default async function handler(request: Request): Promise<Response> {
 
 export async function handlePpr(request: Request): Promise<PprData> {
   const timestamp = new Date().toISOString()
-  const payload: RscPayload = {
+  const rscPayload: RscPayload = {
     root: <Root url={new URL(request.url)} timestamp={timestamp} />,
   }
   const controller = new AbortController()
   const pendingResult = runPrerender(() =>
-    prerender(payload, {
+    prerender(rscPayload, {
       signal: controller.signal,
       onError() {},
     }),
   )
   setTimeout(() => controller.abort(new Error('RSC prerender cutoff')), 0)
   const { prelude } = await pendingResult
-  const ssr = await import.meta.viteRsc.loadModule<
+  const ssrEntryModule = await import.meta.viteRsc.loadModule<
     typeof import('./entry.ssr')
   >('ssr', 'index')
-  const result = await ssr.prerenderHtml(prelude)
+  const result = await ssrEntryModule.prerenderHtml(prelude)
   return {
     html: await new Response(result.prelude).text(),
     postponed: JSON.stringify(result.postponed),
