+++ packages/plugin-rsc/examples/ppr/src/framework/shared.ts
@@ -1,9 +1,3 @@
-import type React from 'react'
-
-export type RscPayload = {
-  root: React.ReactNode
-}
-
 export type PprData = {
   html: string
   postponed: string
